LayoutTests/fast/scrolling/mac/mousewheel-over-scrollbar-expected.txt

asdf

Test scroll over content
PASS overflowScrollEventCount > 0 is true
PASS windowScrollEventCount == 0 is true

Test scroll over border
PASS overflowScrollEventCount > 0 is true
PASS windowScrollEventCount == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
